remove PkgDependency
- implement own version of PkgDependency.builddict() to construct project dependency graph
- IntegrationTests.jl has no dependencies left beside to the standard library
- test implementation with reference implementation from PkgDependency
