+++ Project.toml
@@ -5,11 +5,9 @@ version = "0.1.0-DEV"
 
 [deps]
 Pkg = "44cfe95a-1eb2-52ea-b672-e2afdf69b78f"
-PkgDependency = "9eb5382b-762c-48ca-8139-e736883fe800"
 
 [compat]
-PkgDependency = "0.4"
-julia = "1"
+julia = "1.6"
 
 [extras]
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
